pyproject.toml
[project]
name = "epstein-search"
version = "0.1.0"
description = "Python and CLI client for the DOJ Epstein Library search API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.32.5",
    "pdfplumber>=0.11.0",
    "agno>=1.5.5",
    "openai>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.urls]
Repository = "https://github.com/gwbischof/epstein-search"

[project.scripts]
es = "cli:main"
estl = "timeline:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["client.py", "cli.py", "timeline.py"]
